Moved auth utils to `auth.ts`.
