+++ js/core/src/auth.ts
@@ -14,18 +14,22 @@
  * limitations under the License.
  */
 
+import { AsyncLocalStorage } from 'node:async_hooks';
+import { runInActionRuntimeContext } from './action.js';
+
+const authAsyncLocalStorage = new AsyncLocalStorage<any>();
+
 /**
- * Deletes any properties with `undefined` values in the provided object.
- * Modifies the provided object.
+ * Execute the provided function in the auth context. Call {@link getFlowAuth()} anywhere
+ * within the async call stack to retrieve the auth.
  */
-export function deleteUndefinedProps(obj: any) {
-  for (const prop in obj) {
-    if (obj[prop] === undefined) {
-      delete obj[prop];
-    } else {
-      if (typeof obj[prop] === 'object') {
-        deleteUndefinedProps(obj[prop]);
-      }
-    }
-  }
+export function runWithAuthContext<R>(auth: any, fn: () => R) {
+  return authAsyncLocalStorage.run(auth, () => runInActionRuntimeContext(fn));
+}
+
+/**
+ * Gets the auth object from the current context.
+ */
+export function getFlowAuth(): any {
+  return authAsyncLocalStorage.getStore();
 }
